ecmc.updated param changed to uint64 (asyn >=4.37)

+++ devEcmcSup/main/ecmcGlobalsExtern.h
@@ -50,7 +50,7 @@ extern int                        controllerError;
 extern int                        controllerErrorOld;
 extern int                        controllerReset;
 extern const char                *controllerErrorMsg;
-extern int32_t                    ecmcUpdatedCounter;
+extern uint64_t                   ecmcUpdatedCounter;
 extern int                        asynSkipCyclesFastest;
 extern int                        asynSkipUpdateCounterFastest;
 extern int                        ecTimeoutSeconds;
